feat:complete view dashbaord-admin-patient-doctor

+++ src/main/java/org/mustapha/digitalhospitaljee/service/impl/AdminServiceImpl.java
@@ -3,22 +3,22 @@ package org.mustapha.digitalhospitaljee.service.impl;
 import org.mustapha.digitalhospitaljee.Exceptions.AdminCreationException;
 import org.mustapha.digitalhospitaljee.Exceptions.BusinessException;
 import org.mustapha.digitalhospitaljee.Repository.AdminRepository;
-import org.mustapha.digitalhospitaljee.Repository.impl.AdminRepositoryImpl;
 import org.mustapha.digitalhospitaljee.model.Admin;
+import org.mustapha.digitalhospitaljee.service.AdminService;
 import org.mustapha.digitalhospitaljee.validation.AdminValidator;
 
 import java.util.List;
 
-public class AdminServiceImpl implements AdminRepository{
+public class AdminServiceImpl implements AdminService {
 
-    private  final AdminRepository adminRepository;
+    private final AdminRepository adminRepository;
 
     public AdminServiceImpl(AdminRepository adminRepository){
         this.adminRepository = adminRepository;
     }
 
     @Override
-    public void createAdmin(Admin admin) throws AdminCreationException {
+    public void create(Admin admin) throws AdminCreationException {
         AdminValidator.validate(admin);
         adminRepository.createAdmin(admin);
     }
@@ -31,21 +31,21 @@ public class AdminServiceImpl implements AdminRepository{
 
     @Override
     public void delete(Long id) throws AdminCreationException {
-       if(id <= 0) {
-            throw new BusinessException("the id can not be null for delete");
-       }
+        if(id == null || id <= 0){
+            throw new BusinessException("ID must be positive and not null");
+        }
         adminRepository.delete(id);
     }
 
     @Override
-    public List<Admin> getAdminsList() throws AdminCreationException {
+    public List<Admin> getAllAdmins() throws AdminCreationException {
         return adminRepository.getAdminsList();
     }
 
     @Override
-    public Admin finfById(Long id) throws AdminCreationException {
-        if(id <= 0) {
-            throw new BusinessException("the id can not be null ");
+    public Admin findById(Long id) throws AdminCreationException {
+        if(id == null || id <= 0){
+            throw new BusinessException("ID must be positive and not null");
         }
         return adminRepository.finfById(id);
     }
